chore(consensus): refactor update data validation into UpdateDataValidator
