+++ crates/consensus/genesis/src/updates/operator_fee.rs
@@ -3,8 +3,7 @@
 use alloy_primitives::LogData;
 
 use crate::{
-    OperatorFeeUpdateError, SystemConfig, SystemConfigLog,
-    updates::common::{ValidationError, validate_update_data},
+    OperatorFeeUpdateError, SystemConfig, SystemConfigLog, UpdateDataValidator, ValidationError,
 };
 
 /// The Operator Fee update type.
@@ -31,7 +30,7 @@ impl TryFrom<&SystemConfigLog> for OperatorFeeUpdate {
     fn try_from(log: &SystemConfigLog) -> Result<Self, Self::Error> {
         let LogData { data, .. } = &log.log.data;
 
-        let validated = validate_update_data(data).map_err(|e| match e {
+        let validated = UpdateDataValidator::validate(data).map_err(|e| match e {
             ValidationError::InvalidDataLen(_expected, actual) => {
                 OperatorFeeUpdateError::InvalidDataLen(actual)
             }
