+++ crates/consensus/genesis/src/updates/signer.rs
@@ -3,10 +3,7 @@
 use alloy_primitives::{Address, LogData};
 use alloy_sol_types::{SolType, sol};
 
-use crate::{
-    SystemConfigLog, UnsafeBlockSignerUpdateError,
-    updates::common::{ValidationError, validate_update_data},
-};
+use crate::{SystemConfigLog, UnsafeBlockSignerUpdateError, UpdateDataValidator, ValidationError};
 
 /// The unsafe block signer update type.
 #[derive(Debug, Clone, Hash, PartialEq, Eq)]
@@ -22,7 +19,7 @@ impl TryFrom<&SystemConfigLog> for UnsafeBlockSignerUpdate {
     fn try_from(log: &SystemConfigLog) -> Result<Self, Self::Error> {
         let LogData { data, .. } = &log.log.data;
 
-        let validated = validate_update_data(data).map_err(|e| match e {
+        let validated = UpdateDataValidator::validate(data).map_err(|e| match e {
             ValidationError::InvalidDataLen(_expected, actual) => {
                 UnsafeBlockSignerUpdateError::InvalidDataLen(actual)
             }
